chore: refactor vrf structures
