+++ libraries/types/vote/CMakeLists.txt
@@ -1,5 +1,5 @@
-set(HEADERS include/vote/vote.hpp)
-set(SOURCES src/vote.cpp)
+set(HEADERS include/vote/vote.hpp include/vote/vrf_sortition.hpp)
+set(SOURCES src/vote.cpp src/vrf_sortition.cpp)
 
 add_library(vote ${SOURCES} ${HEADERS})
 target_include_directories(vote PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
