+++ tests/crypto_test.cpp
@@ -235,10 +235,10 @@ TEST_F(CryptoTest, DISABLED_compute_vdf_solution_cost_time) {
 
   SortitionParams sortition_params(32768, 29184, 16, 21, 22, 100);
   VdfSortition vdf(sortition_params, sk, getRlpBytes(level));
-  std::cout << "output " << vdf.output << std::endl;
+  std::cout << "output " << vdf.output_ << std::endl;
   int i = 0;
-  for (; i < vdf.output.size; i++) {
-    std::cout << vdf.output[i] << ", " << vdf.output.hex()[i] << ", " << uint16_t(vdf.output[i]) << std::endl;
+  for (; i < vdf.output_.size; i++) {
+    std::cout << vdf.output_[i] << ", " << vdf.output_.hex()[i] << ", " << uint16_t(vdf.output_[i]) << std::endl;
   }
   std::cout << "size: " << i << std::endl;
 }
@@ -357,7 +357,8 @@ TEST_F(CryptoTest, leader_selection) {
   const uint64_t high_stake_nodes_num = 2;
   const uint64_t high_stake_nodes_power = std::rand() % 2000;
   const uint64_t low_stake_nodes_power = std::rand() % 50;
-  const auto valid_sortition_players = high_stake_nodes_num * high_stake_nodes_power + low_stake_nodes_num * low_stake_nodes_power;
+  const auto valid_sortition_players =
+      high_stake_nodes_num * high_stake_nodes_power + low_stake_nodes_num * low_stake_nodes_power;
 
   for (uint64_t i = 0; i < high_stake_nodes_num; i++) {
     high_stake_nodes.emplace(i, getVrfKeyPair().second);
@@ -366,40 +367,31 @@ TEST_F(CryptoTest, leader_selection) {
     low_stake_nodes.emplace(i + high_stake_nodes_num, getVrfKeyPair().second);
   }
 
+  const auto selector = [&](auto& outputs, const auto& msg, const auto& nodes, auto node_power) {
+    for (const auto& n : nodes) {
+      VrfPbftSortition sortition(n.second, msg);
+      HashableVrf hash(sortition.output_, n.second, 0);
+      if (auto stake = VrfPbftSortition::getBinominalDistribution(node_power, valid_sortition_players, committee_size,
+                                                                  hash.getHash())) {
+        hash.iter = 1;
+        auto lowest = hash.getHash();
+        for (uint64_t j = 2; j <= stake; j++) {
+          hash.iter = j;
+          auto tmp = hash.getHash();
+          if (tmp < lowest) lowest = std::move(tmp);
+        }
+        outputs.emplace(n.first, lowest);
+      }
+    }
+  };
+
   for (uint64_t i = 0; i < rounds; i++) {
     const VrfPbftMsg msg(propose_vote_type, i, 1);
     std::unordered_map<uint64_t, dev::h256> outputs;
-    for (const auto& n : high_stake_nodes) {
-      VrfPbftSortition sortition(n.second, msg);
-      HashableVrf hash(sortition.output, n.second, 0);
-      if (auto stake = VrfPbftSortition::getBinominalDistribution(
-              high_stake_nodes_power, valid_sortition_players, committee_size, hash.getHash())) {
-        hash.iter = 1;
-        auto lowest = hash.getHash();
-        for (uint64_t j = 2; j <= stake; j++) {
-          hash.iter = j;
-          const auto tmp = hash.getHash();
-          if (tmp < lowest) lowest = std::move(tmp);
-        }
-        outputs.emplace(n.first, lowest);
-      }
-    }
 
-    for (const auto& n : low_stake_nodes) {
-      VrfPbftSortition sortition(n.second, msg);
-      HashableVrf hash(sortition.output, n.second, 0);
-      if (auto stake = VrfPbftSortition::getBinominalDistribution(
-              low_stake_nodes_power, valid_sortition_players, committee_size, hash.getHash())) {
-        hash.iter = 1;
-        auto lowest = hash.getHash();
-        for (uint64_t j = 2; j <= stake; j++) {
-          hash.iter = j;
-          const auto tmp = hash.getHash();
-          if (tmp < lowest) lowest = std::move(tmp);
-        }
-        outputs.emplace(n.first, lowest);
-      }
-    }
+    selector(outputs, msg, high_stake_nodes, high_stake_nodes_power);
+    selector(outputs, msg, low_stake_nodes, low_stake_nodes_power);
+
     // every round there needs to be a leader
     EXPECT_TRUE(outputs.size());
     const auto leader = *std::min_element(outputs.begin(), outputs.end(),
@@ -410,14 +402,14 @@ TEST_F(CryptoTest, leader_selection) {
   uint64_t high_stake_nodes_blocks = 0;
   for (const auto& node : block_produced) {
     if (node.first == high_stake_nodes_num) break;
-     high_stake_nodes_blocks += node.second;
+    high_stake_nodes_blocks += node.second;
   }
 
   const auto stake_ratio = high_stake_nodes_power * high_stake_nodes_num * 100 / valid_sortition_players;
   const auto blocks_ratio = high_stake_nodes_blocks * 100 / rounds;
   std::cout << "Stake ratio: " << stake_ratio << " Blocks ratio:" << blocks_ratio << std::endl;
-  const auto diff = (stake_ratio > blocks_ratio) ? (stake_ratio-blocks_ratio) : (blocks_ratio - stake_ratio);
-  //maximal difference is 3%
+  const auto diff = (stake_ratio > blocks_ratio) ? (stake_ratio - blocks_ratio) : (blocks_ratio - stake_ratio);
+  // maximal difference is 3%
   EXPECT_LE(diff, 3);
 }
